backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.7
langchain-openai==0.2.9
langgraph==0.2.53
python-dotenv==1.0.1
pydantic==2.9.2
pymongo==4.11.0
langgraph-checkpoint-mongodb==0.1.3
langchain-groq==0.2.1
motor==3.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
certifi
python-multipart==0.0.12
email-validator==2.2.0
pdfplumber
pytesseract
Pillow
